data type changed

+++ AddOptimization.Contracts/Dto/QuoteSummaryDto.cs
@@ -6,11 +6,11 @@ namespace AddOptimization.Contracts.Dto
     public class QuoteSummaryDto
     {
         public string Name { get; set; }
-        public int Quantity { get; set; }
-        public int Vat { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Vat { get; set; }
         public decimal UnitPrice { get; set; }
-        public int TotalPriceExcVat { get; set; }
-        public int TotalPriceIncVat { get; set; }
+        public decimal TotalPriceExcVat { get; set; }
+        public decimal TotalPriceIncVat { get; set; }
         public Guid QuoteId { get; set; }
     }
 }
