+++ AddOptimization.Data/Entities/QuoteSummary.cs
@@ -9,11 +9,11 @@ namespace AddOptimization.Data.Entities
         public Guid Id { get; set; }
         public Guid QuoteId { get; set; }
         public string Name { get; set; }
-        public int Quantity { get; set; }
-        public int Vat { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Vat { get; set; }
         public decimal UnitPrice { get; set; }
-        public int TotalPriceExcVat { get; set; }
-        public int TotalPriceIncVat { get; set; }
+        public decimal TotalPriceExcVat { get; set; }
+        public decimal TotalPriceIncVat { get; set; }
 
         [ForeignKey(nameof(QuoteId))]
         public virtual Quote Quote { get; set; }
